Extract signal handler setup and configuration into separate file
This will make it easier to reuse this code in upcoming tests.

This code already had test coverage in the
`Collectors::CpuAndWallTimeWorker` specs, and I decided to not separate
out the tests.
